Trying out suggested improvements to the XML unit tests.
