+++ sal_interfaces/Script/Script_Commands.xml
@@ -135,7 +135,7 @@
     <Subsystem>Script</Subsystem>
     <Version>1.0.0</Version>
     <Author></Author>
-    <EFDB_Topic>Script_command_stop</EFDB_Topic>
+    <EFDB_Topic>Script_command_sTop</EFDB_Topic>
     <Alias>stop</Alias>
     <Device></Device>
     <Property></Property>
